+++ tutorial/unicode.py
@@ -8,8 +8,8 @@ pdf.add_page()
 # Add a DejaVu Unicode font (uses UTF-8)
 # Supports more than 200 languages. For a coverage status see:
 # http://dejavu.svn.sourceforge.net/viewvc/dejavu/trunk/dejavu-fonts/langcover.txt
-pdf.add_font("DejaVu", fname="DejaVuSansCondensed.ttf")
-pdf.set_font("DejaVu", size=14)
+pdf.add_font(fname="DejaVuSansCondensed.ttf")
+pdf.set_font("DejaVuSansCondensed", size=14)
 
 text = """
 English: Hello World
@@ -30,14 +30,14 @@ for txt in text.split("\n"):
 # Supports: Bengali, Devanagari, Gujarati,
 #           Gurmukhi (including the variants for Punjabi)
 #           Kannada, Malayalam, Oriya, Tamil, Telugu, Tibetan
-pdf.add_font("gargi", fname="gargi.ttf")
+pdf.add_font(fname="gargi.ttf")
 pdf.set_font("gargi", size=14)
 pdf.write(8, "Hindi: नमस्ते दुनिया")
 pdf.ln(20)
 
 # Add a AR PL New Sung Unicode font (uses UTF-8)
 # The Open Source Chinese Font (also supports other east Asian languages)
-pdf.add_font("fireflysung", fname="fireflysung.ttf")
+pdf.add_font(fname="fireflysung.ttf")
 pdf.set_font("fireflysung", size=14)
 pdf.write(8, "Chinese: 你好世界\n")
 pdf.write(8, "Japanese: こんにちは世界\n")
@@ -46,14 +46,14 @@ pdf.ln(10)
 # Add a Alee Unicode font (uses UTF-8)
 # General purpose Hangul truetype fonts that contain Korean syllable
 # and Latin9 (iso8859-15) characters.
-pdf.add_font("eunjin", fname="Eunjin.ttf")
-pdf.set_font("eunjin", size=14)
+pdf.add_font(fname="Eunjin.ttf")
+pdf.set_font("Eunjin", size=14)
 pdf.write(8, "Korean: 안녕하세요")
 pdf.ln(20)
 
 # Add a Fonts-TLWG (formerly ThaiFonts-Scalable) (uses UTF-8)
-pdf.add_font("waree", fname="Waree.ttf")
-pdf.set_font("waree", size=14)
+pdf.add_font(fname="Waree.ttf")
+pdf.set_font("Waree", size=14)
 pdf.write(8, "Thai: สวัสดีชาวโลก")
 pdf.ln(20)